Title: 낮은 지점들, Time: 124ms, Memory: 2MB, Status: Passed - CodeTree

+++ 240116/낮은 지점들/lowest-points.cpp
@@ -31,7 +31,7 @@ int main()
 			um[a] = min(um[a], b);
 		}
 	}
-	int answer = 0;
+	long long answer = 0;
 	for (auto i : um)
 	{
 		answer += i.second;
